Implement support for adding collaborators to the repo
This change adds a new action `joinProjectAsCollaborator` in the `github.ts` file to allow users to join a project as a collaborator. The action checks if the user is already a member of the project and if not, sends an invitation to the user to join the project's GitHub repository as a collaborator with 'push' permissions. If the invitation is already sent or the user is already a collaborator, the action returns the appropriate success message. The action also handles the case where the user is not authenticated or the project is not found.

In the `members.ts` file, two internal queries `getMember` and `addMember` are added to retrieve information about members and add a new member to a project, respectively. The `getMember` query searches for a member by projectId and userId, while the `addMember` mutation adds a new member with the provided projectId and userId.

In the `projects.ts` file, an internal query `getWithCode` is added to retrieve project information using an inviteCode. Similarly, a query `projectIdForCode` is added to retrieve the projectId associated with an inviteCode.

Some changes to the UI are made in `InvitePeopleModal.tsx` to display a note informing users that anyone who joins using the invite link will be added as a collaborator with 'push' permissions.

In the `JoinPage.tsx` file, a button to accept the invitation is added an the onAccept function calls the joinProject action and redirects the user to the project if successfull